+++ push-20160801/alibabacloud_push20160801/client.py
@@ -1191,6 +1191,10 @@ class Client(OpenApiClient):
             query['AndroidXiaoMiNotifyTitle'] = request.android_xiao_mi_notify_title
         if not DaraCore.is_null(request.android_xiaomi_big_picture_url):
             query['AndroidXiaomiBigPictureUrl'] = request.android_xiaomi_big_picture_url
+        if not DaraCore.is_null(request.android_xiaomi_focus_param):
+            query['AndroidXiaomiFocusParam'] = request.android_xiaomi_focus_param
+        if not DaraCore.is_null(request.android_xiaomi_focus_pics):
+            query['AndroidXiaomiFocusPics'] = request.android_xiaomi_focus_pics
         if not DaraCore.is_null(request.android_xiaomi_image_url):
             query['AndroidXiaomiImageUrl'] = request.android_xiaomi_image_url
         if not DaraCore.is_null(request.app_key):
@@ -1457,6 +1461,10 @@ class Client(OpenApiClient):
             query['AndroidXiaoMiNotifyTitle'] = request.android_xiao_mi_notify_title
         if not DaraCore.is_null(request.android_xiaomi_big_picture_url):
             query['AndroidXiaomiBigPictureUrl'] = request.android_xiaomi_big_picture_url
+        if not DaraCore.is_null(request.android_xiaomi_focus_param):
+            query['AndroidXiaomiFocusParam'] = request.android_xiaomi_focus_param
+        if not DaraCore.is_null(request.android_xiaomi_focus_pics):
+            query['AndroidXiaomiFocusPics'] = request.android_xiaomi_focus_pics
         if not DaraCore.is_null(request.android_xiaomi_image_url):
             query['AndroidXiaomiImageUrl'] = request.android_xiaomi_image_url
         if not DaraCore.is_null(request.app_key):