+++ push-20160801/alibabacloud_push20160801/models/_mass_push_request.py
@@ -117,6 +117,8 @@ class MassPushRequestPushTask(DaraModel):
         android_xiao_mi_notify_body: str = None,
         android_xiao_mi_notify_title: str = None,
         android_xiaomi_big_picture_url: str = None,
+        android_xiaomi_focus_param: str = None,
+        android_xiaomi_focus_pics: str = None,
         android_xiaomi_image_url: str = None,
         body: str = None,
         device_type: str = None,
@@ -228,6 +230,8 @@ class MassPushRequestPushTask(DaraModel):
         self.android_xiao_mi_notify_body = android_xiao_mi_notify_body
         self.android_xiao_mi_notify_title = android_xiao_mi_notify_title
         self.android_xiaomi_big_picture_url = android_xiaomi_big_picture_url
+        self.android_xiaomi_focus_param = android_xiaomi_focus_param
+        self.android_xiaomi_focus_pics = android_xiaomi_focus_pics
         self.android_xiaomi_image_url = android_xiaomi_image_url
         self.body = body
         # This parameter is required.
@@ -458,6 +462,12 @@ class MassPushRequestPushTask(DaraModel):
 
         if self.android_xiaomi_big_picture_url is not None:
             result['AndroidXiaomiBigPictureUrl'] = self.android_xiaomi_big_picture_url
+
+        if self.android_xiaomi_focus_param is not None:
+            result['AndroidXiaomiFocusParam'] = self.android_xiaomi_focus_param
+
+        if self.android_xiaomi_focus_pics is not None:
+            result['AndroidXiaomiFocusPics'] = self.android_xiaomi_focus_pics
 
         if self.android_xiaomi_image_url is not None:
             result['AndroidXiaomiImageUrl'] = self.android_xiaomi_image_url
@@ -792,6 +802,12 @@ class MassPushRequestPushTask(DaraModel):
 
         if m.get('AndroidXiaomiBigPictureUrl') is not None:
             self.android_xiaomi_big_picture_url = m.get('AndroidXiaomiBigPictureUrl')
+
+        if m.get('AndroidXiaomiFocusParam') is not None:
+            self.android_xiaomi_focus_param = m.get('AndroidXiaomiFocusParam')
+
+        if m.get('AndroidXiaomiFocusPics') is not None:
+            self.android_xiaomi_focus_pics = m.get('AndroidXiaomiFocusPics')
 
         if m.get('AndroidXiaomiImageUrl') is not None:
             self.android_xiaomi_image_url = m.get('AndroidXiaomiImageUrl')